Fix function move operation
The move operation in `vt_proc` had a substantial bug that surprisingly
only surfaced with optimized x68 gcc builds: the move-constructed
instance would immediately be destroyed again, while the intended
behaviour was to destroy the instance that was moved from.

This likely didn't surface because the storage was unaffected by the
destructor call most of the time. However, from looking at the code
generated by gcc, it seems it's optimizing away the move construction
completely.

In any case, this change corrects the behaviour. The argument order
to `vt_proc` has been swapped to ensure the most compact code is
generated. Interestingly, code generated by arm-gcc is actually smaller
with the new, correct implementation.